+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `About` window with versioning info
 * Add `Android Projects App` product
+* Add `DocumentServer Installation` product
 
 ### Fixes